feat(tt-rss) reimplement with upstream container images (#17293)
**Description**
It appears that the k8s-at-home project which this app was based on is
unmaintained, and the container has stopped running successfully for
myself and others. This pull request moves the tt-rss app back to dev
train, and includes an initial attempt at migrating from k8s-at-home to
the tt-rss author's docker images.
<!--
Please include a summary of the change and which issue is fixed. Please
also include relevant motivation and context. List any dependencies that
are required for this change.
-->

⚒️ Approaches fix for  #8647 <!--(issue)-->

// EDIT: Stavros
Closes #8647

**⚙️ Type of change**

- [ ] ⚙️ Feature/App addition
- [x] 🪛 Bugfix
- [ ] ⚠️ Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] 🔃 Refactor of current code

**🧪 How Has This Been Tested?**
Tested with helm+kubernetes on local Docker Desktop instance. Deploys
successfully with a local.yaml file which sets LoadBalancer and the Self
URL variable:
```
$ cat local.yaml 
service:
  main:
    type: LoadBalancer
workload:
  main:
    podSpec:
      containers:
        main:
          env:
            TTRSS_SELF_URL_PATH: "http://localhost:10104/tt-rss/"
```
<!--
Please describe the tests that you ran to verify your changes. Provide
instructions so we can reproduce. Please also list any relevant details
for your test configuration
-->

**📃 Notes:**
I have not tried implementing updated TrueNAS questions yet, or made
updates to the app documentation. The app does not provide semantic
versioning that I have found, versions arbitrarily bumped.
<!-- Please enter any other relevant information here -->

**✔️ Checklist:**

- [x] ⚖️ My code follows the style guidelines of this project
- [x] 👀 I have performed a self-review of my own code
- [ ] #️⃣ I have commented my code, particularly in hard-to-understand
areas
- [ ] 📄 I have made corresponding changes to the documentation
- [ ] ⚠️ My changes generate no new warnings
- [ ] 🧪 I have added tests to this description that prove my fix is
effective or that my feature works
- [ ] ⬆️ I increased versions for any altered app according to semantic
versioning

**➕ App addition**

If this PR is an app addition please make sure you have done the
following.

- [x] 🪞 I have opened a PR on
[truecharts/containers](https://github.com/truecharts/containers) adding
the container to TrueCharts mirror repo.
- [x] 🖼️ I have added an icon in the Chart's root directory called
`icon.png`

---

_Please don't blindly check all the boxes. Read them and only check
those that apply.
Those checkboxes are there for the reviewer to see what is this all
about and
the status of this PR with a quick glance._

---------

Signed-off-by: Christopher <bigbero@gmail.com>
Signed-off-by: Stavros Kois <47820033+stavros-k@users.noreply.github.com>
Co-authored-by: Stavros Kois <47820033+stavros-k@users.noreply.github.com>